refactor: Split GraphNode logic from general graph logic in core
