+++ crates/gantz_core/src/lib.rs
@@ -1,5 +1,4 @@
 pub use edge::Edge;
-pub use graph::Graph;
 pub use node::Node;
 pub use steel;
